+++ adminforth/index.ts
@@ -21,6 +21,7 @@ import {
   IAdminForthDataSourceConnectorBase,
   IWebSocketBroker,
   HttpExtra,
+  BeforeCreateSaveFunction,
 } from './types/Back.js';
 
 import {
@@ -113,6 +114,11 @@ class AdminForth implements IAdminForth {
     };
 
     console.log(`🚀 AdminForth v${ADMINFORTH_VERSION} starting up`)
+  }
+
+  async tr(this, msg: string, category: string = 'default', lang: string = 'en'): Promise<string> {
+    // stub function to make a translation
+    return msg;
   }
 
   activatePlugins() {
@@ -295,7 +301,7 @@ class AdminForth implements IAdminForth {
   ): Promise<{ error?: string, createdRecord?: any }> {
 
     // execute hook if needed
-    for (const hook of listify(resource.hooks?.create?.beforeSave as BeforeSaveFunction[])) {
+    for (const hook of listify(resource.hooks?.create?.beforeSave as BeforeCreateSaveFunction[])) {
       console.log('🪲 Hook beforeSave', hook);
       const resp = await hook({ 
         recordId: undefined,