+++ dev-demo/migrations/20241214192221_add_source_and_cat_to_tr/migration.sql
@@ -1,0 +1,5 @@
+-- AlterTable
+ALTER TABLE "translations" ADD COLUMN "category" TEXT;
+
+-- CreateIndex
+CREATE INDEX "translations_category_idx" ON "translations"("category");