+++ dev-demo/migrations/20241214200001_tr_complete/migration.sql
@@ -1,0 +1,2 @@
+-- AlterTable
+ALTER TABLE "translations" ADD COLUMN "completed" TEXT;