+++ dev-demo/migrations/20241214211735_improve_tr_incex/migration.sql
@@ -1,0 +1,2 @@
+-- CreateIndex
+CREATE INDEX "translations_category_idx" ON "translations"("category");